Read fixed-width files quickly.

Experimental replacement for utils::read.fwf that runs much faster. However, it is much less flexible than utils::read.fwf.

Usage

fast.read.fwf(
  file,
  widths,
  col.names = NULL,
  colClasses = NA,
  tz = "",
  dec = ".",
  ...
)

Arguments

file

Character vector: name of file

widths

Numeric vector: column widths. Negative numbers mean "skip this many columns". Use an NA as the final element if there are likely to be extra characters at the end of each row after the last one that you're interested in.

col.names

names for the columns that are NOT skipped

colClasses

can be used to control type conversion; see read.table(). It is an optional vector whose names must be part of col.names. There is one extension of the read.table() rules: a colClasses string starting POSIXct will trigger automatic conversion to POSIXct, using the rest of the string as the format specifier.

tz

used in auto-conversion to POSIXct() when colClasses is set

dec

the character to be assumed for decimal points. Passed to utils::type.convert()

...

ignored