Reads a triple-s XML (asc) data file, as specified by the triple-s XML standard.

This function reads and parses a .sss XML metadata file as well as its associated .asc data file. The sss standard defines a standard survey structure.

read.sss(sssFilename, ascFilename = guess_asc_filename(sssFilename), sep = "_")

Arguments

sssFilename
Character string: name of .sss file containing the survey metadata

ascFilename
Character string: name of .asc (or .csv) file containing survey data. If this is not provided, guesses the filename using

sep
Character vector defining the string that separates question and subquestion labels, e.g. c("Q_1", "Q_2")

Value

A data frame with one element (column) for each variable in the data set.

The resulting data.frame contains several attributes:

variable.labels: a named list of value labels with one element per variable, either NULL or a named character vector

label.table: a named list with one element per question. Every element is a named character string that contains the label codes for that question.

References

http://www.triple-s.org/

See also

Examples

example <- system.file("sampledata/sample-1.sss", package = "sss")

# read.sss() automatically guesses the data file name
read.sss(system.file("sampledata/sample-1.sss", package = "sss"))

#> Reading SSS metadata

#> Reading SSS data

#>                               Q1 Q2_1 Q2_2 Q2_3 Q2_4 Q2_5 Q2_6 Q2_7 Q2_8 Q2_9
#> 1 Visited before within the year    1    0    1    0    0    0    0    0    1
#> 2            Visited before that    0    1    0    0    0    0    0    0    0
#> 3 Visited before within the year    1    0    0    1    0    0    0    0    1
#>                               Q3              Q4_1               Q4_2  Q5    Q6
#> 1 Amusement Park                   Sherwood Forest              Other  12  TRUE
#> 2                                Nottingham Castle               <NA> 999  TRUE
#> 3 "Marco's" Restaurant                       Other "Maid Marion" Cafe  58 FALSE
#>                       Q7 Q99
#> 1        Within 3 months 1.4
#> 2 More than 1 years time 0.9
#> 3                   <NA> 0.7

read.sss(system.file("sampledata/sample-2.sss", package = "sss"))

#> Reading SSS metadata

#> Reading SSS data

#>                             MIProRspId AltIdString AltIdNumeric  AdateF1
#> 1 c530fd2a-17dd-4232-a38d-fb786d85eb3f                        0 20150805
#>   AtimeF1 AacN1.1 AstudyF1 AaltidN1.1 AlangN1.1   ASysInfo1M1.1 ASysInfo1M1.2
#> 1  121936                0                      AppleMAC-Safari       Windows
#>   ASysInfo1M1.3
#> 1              
#>                                                                                                    ASysInfo1M1.4
#> 1 Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML; like Gecko) Chrome/44.0.2403.125 Safari/537.36
#>   ASysInfo1M1.5               AA01N1 ASysInfo2_1M1.1 ASysInfo2_1M1.2
#> 1   86.6.118.43 <text>English</text>          Chrome   44.0.2403.125
#>   ASysInfo2_1M1.3 ASysInfo2_1M1.4
#> 1              44        Netscape
#>                                                                                                  ASysInfo2_1M1.5
#> 1 Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/44.0.2403.125 Safari/537.36
#>   ASysInfo2_1M1.6 ASysInfo2_1M1.7 ASysInfo2_1M1.8 ASysInfo2_1M1.9
#> 1           Win32     1920 x 1080    24-bit color           false
#>               AA02N1                 AA03N1             AA04N1           AA05N1
#> 1 <text>Qatar</text> <text>Umm Salal</text> <text>Urban</text> <text>Yes</text>
#>   AA06N1 AA07N1            AA08N1      ADOBDayN1      ADOBMonthN1
#> 1   <NA>   <NA> <text>Male</text> <text>5</text> <text>Jul</text>
#>          ADOBYearN1     AA09M1.1         AA10N1         AA10N2         AA10N3
#> 1 <text>1991</text> 5  Jul  1991 <text>4</text> <text>3</text> <text>2</text>
#>           AA10N4         AA10N5            AA10N6           AA11N1
#> 1 <text>1</text> <text>2</text> <text>None</text> <text>Yes</text>
#>                                           AA12N1 AA13N1
#> 1 <text>Manual worker / builder / cleaner</text>   <NA>
#>                                   AA14N1      AA15M1   AA16M1
#> 1 <text>$1,001 - $2,000 per month</text> 00100100110 11100000
#>                     AA17N1                         AA17N2
#> 1 <text>Once a week</text> <text>2-3 times a month</text>
#>                          AA17N3                        AA17N4
#> 1 <text>4-5 times a week</text> <text>2-3 times a week</text>
#>                          AA17N5                   AA17N6
#> 1 <text>2-3 times a week</text> <text>Once a week</text>
#>                           AA17N7                    AA17N8
#> 1 <text>2-3 times a month</text> <text>Once a month</text>
#>                     AA17N9                        AA17N10    AA18M1 AA19M1.1
#> 1 <text>Once a week</text> <text>2-3 times a month</text> 010111000      Pat
#>   AA19M1.2               AA19M1.3 AA19M1.4             AStatusN1 AEndTF1
#> 1   Molloy pat.molloy@outlook.com   112233 <text>Complete</text>  122113
#>   ADurationF1
#> 1          97

read.sss(system.file("sampledata/sample-3.sss", package = "sss"))

#> Error: Unable to find guessed data file at /Users/runner/work/_temp/Library/sss/sampledata/sample-3.asc