Reads a triple-s XML (asc) data file, as specified by the triple-s XML standard.

This function reads and parses a .sss XML metadata file as well as its associated .asc data file. The sss standard defines a standard survey structure.

read.sss(sssFilename, ascFilename, sep = "_")

Arguments

sssFilename
Character string: name of .sss file containing the survey metadata

ascFilename
Character string: name of .asc file containing survey data

sep
Character vector defining the string that separates question and subquestion labels, e.g. c("Q_1", "Q_2")

Value

A data frame with one element (column) for each variable in the data set.

The data.frame contains several attributes:

variable.labels: a named list of value labels with one element per variable, either NULL or a named character vector

label.table: a named list with one element per question. Every element is a named character strings, that contains the label codes for that question.

References

http://www.triple-s.org/

See also

Examples

sampleRoot <- system.file("sampledata", package = "sss")
filenameSSS <- file.path(sampleRoot, "sample-1.sss")

read.sss(filenameSSS)

#> Reading SSS metadata

#> Reading SSS data

#>                               Q1 Q2_1 Q2_2 Q2_3 Q2_4 Q2_5 Q2_6 Q2_7 Q2_8 Q2_9
#> 1 Visited before within the year    1    0    1    0    0    0    0    0    1
#> 2            Visited before that    0    1    0    0    0    0    0    0    0
#> 3 Visited before within the year    1    0    0    1    0    0    0    0    1
#>                               Q3              Q4_1               Q4_2  Q5    Q6
#> 1 Amusement Park                   Sherwood Forest              Other  12  TRUE
#> 2                                Nottingham Castle               <NA> 999  TRUE
#> 3 "Marco's" Restaurant                       Other "Maid Marion" Cafe  58 FALSE
#>                       Q7 Q99
#> 1        Within 3 months 1.4
#> 2 More than 1 years time 0.9
#> 3                   <NA> 0.7